\name{electricity}
\alias{electricity}
\docType{data}
\title{
  Hourly Electricity Prices in the Spanish Market 
}
\description{
  Partial realizations of time series of hourly electricity prices (Cent/kWh) in the Spanish market.
}
\usage{data(electricity)}
\format{
  A matrix with 365 observations of the hourly electricity cost readings.
  
}
\references{
Montero, P and Vilar, J.A. (2014) \emph{TSclust: An R Package for Time Series Clustering.}  Journal of Statistical Software, 62(1), 1-43. \url{http://www.jstatsoft.org/v62/i01/.}
}
\details{
The dataset consists of hourly electricity prices in the Spanish market
during weekdays of the period December 31, 2007 - May 25, 2009.
}
