\name{interest.rates}
\alias{interest.rates}
\docType{data}
\title{
Long-Term Interest Rates from 1995 to 2012
}
\description{
  Partial realizations of time series of long-term interest rates (10-year bonds) for several countries. 
}
\usage{data(interest.rates)}
\format{
  A \code{ts} object with 215 observations of the monthly long-term interest rates (10-year bonds) from January 1995 to November 2012 of several countries.
}


