\name{electricity}
\alias{electricity}
\docType{data}
\title{
  Hourly Electricity Prices in the Spanish Market 
}
\description{
  Partial realizations of time series of hourly electricity prices (Cent/kWh) in the Spanish market.
}
\usage{data(electricity)}
\format{
  A matrix with 365 observations of the hourly electricity cost readings.
  
}
\details{
The dataset consists of hourly electricity prices in the Spanish market
during weekdays of the period December 31, 2007 - May 25, 2009. Available at \url{http://www.omel.es}.

}
