\name{interest.rates}
\alias{interest.rates}
\docType{data}
\title{
Long-Term Interest Rates from 1995 to 2012
}
\description{
  Partial realizations of time series of long-term interest rates (10-year bonds) for several countries. 
}
\references{
Montero, P and Vilar, J.A. (2014) \emph{TSclust: An R Package for Time Series Clustering.}  Journal of Statistical Software, 62(1), 1-43. \doi{doi:10.18637/jss.v062.i01}
}
\usage{data(interest.rates)}
\format{
  A \code{ts} object with 215 observations of the monthly long-term interest rates (10-year bonds) from January 1995 to November 2012 of several countries.
}


